Cambiado forma de registrar porcentaje de progreso y el calculo de indicadores de los indicadores E1I2 y E3I2

+++ app/controllers/cor1440_gen/tiposindicador_controller.rb
@@ -12,7 +12,10 @@ module Cor1440Gen
         "medircon",
         "esptipometa",
         "campotind",
-        "pprogtind",
+        "desc20",
+        "desc40",
+        "desc80",
+        "desc100",
         "espvaloresomision",
         "espvalidaciones",
         "espfuncionmedir" ]
@@ -27,10 +30,13 @@ module Cor1440Gen
     end
 
     def update
-      params[:tipoindicador][:campotind_attributes].each do |p| 
-        if p[1]['_destroy'] == "1" && p[1][:id].to_i > 0
-          ::EfectoValorcampotind.where(valorcampotind_id: p[1][:id]).destroy_all
-          Cor1440Gen::Valorcampotind.where(campotind_id: p[1][:id]).destroy_all
+      byebug
+      if params[:tipoindicador] && params[:tipoindicador][:campotind_attributes]
+        params[:tipoindicador][:campotind_attributes].each do |p| 
+          if p[1]['_destroy'] == "1" && p[1][:id].to_i > 0
+            ::EfectoValorcampotind.where(valorcampotind_id: p[1][:id]).destroy_all
+            Cor1440Gen::Valorcampotind.where(campotind_id: p[1][:id]).destroy_all
+          end
         end
       end
       update_gen
